requirements.txt
upstox-python
python-dotenv
schedule
